continuación

+++ java/entities/User.java
@@ -35,7 +35,7 @@ public class User implements Serializable {
   private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String username;
     
@@ -55,14 +55,15 @@ public class User implements Serializable {
     private UserPrivilege privilege;
     
     private String password;
-    
-     public Integer getId() {
+
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
+
     public String getUsername() {
         return username;
     }
